Database code
This version uses the mongo db library mixed with an extra 3rd party
library called mgo to create a database that I can use to add entries
to my database. If the database could get on a server then people could
actively be changing the status of test. However, this is not finished
and the system that organizes the entries into the ui is unfinished