Adding to readme.
This will be the final commit of July and the internship. I cleaned up
everything and commented the go code. I got rid of the unnecessary json
code and the plaintext files used for the simplistic web development
that this whole project was built off of. I also deleted the extra html
documents excluding the one based off of bootstrap example integrated
into the go file and included as a separate file in the fold. I renamed
wiki.go to dashboard.go as that was the original project. This version
does integrate mongo.db